Adressing issues
