+++ src/utils/inc/brainflow_constants.h
@@ -89,3 +89,14 @@ enum class BrainFlowClassifiers : int
     REGRESSION = 0,
     KNN = 1
 };
+/// LogLevels enum to store all possible log levels
+enum class LogLevels : int
+{
+    LEVEL_TRACE = 0,    /// TRACE
+    LEVEL_DEBUG = 1,    /// DEBUG
+    LEVEL_INFO = 2,     /// INFO
+    LEVEL_WARN = 3,     /// WARN
+    LEVEL_ERROR = 4,    /// ERROR
+    LEVEL_CRITICAL = 5, /// CRITICAL
+    LEVEL_OFF = 6       // OFF
+};
